prince-or-not/requirements.txt
streamlit
opencv-python
Pillow
numpy
torch
facenet-pytorch
deepface 